Refactored PPM protocols
git-svn-id: http://source.eict.vpdc/svn/labo/protocols/trunk@38842 0d108963-8a47-0410-809b-ebd9fb3b3407

+++ upl/com.energyict.mdc.upl.impl/protocolimpl/src/main/java/com/energyict/protocolimpl/iec1107/ppmi1/parser/NumberAssembler.java
@@ -26,13 +26,13 @@ class NumberAssembler implements Assembler {
 	public void workOn(Assembly ta) throws IOException {
 
 		Day day = (Day) ta.getTarget();
-		Byte bte = ((Byte) ta.pop()).byteValue();
+		byte bte = ((Byte)ta.pop()).byteValue();
 
 		if (day == null) {
 			return;
 		}
 
-		getVal()[this.byteNr] = bte.byteValue();
+		getVal()[this.byteNr] = bte;
 		this.byteNr++;
 
 		if (this.byteNr != (getProfileParser().getNrOfChannels() * BYTES_PER_VALUE + BYTES_PER_STATUS)) {
